Reset TableViewController to start over
